Add evince version check, fixes #1133
